backend/requirements_trip_planner.txt
# Requirements for Enhanced Interactive Trip Planner
# Install with: pip install -r requirements_trip_planner.txt

# For loading environment variables from .env file
python-dotenv>=1.0.0

# For making HTTP requests to Google Maps API
requests>=2.28.0

# Standard library modules (no installation needed):
# - json
# - os
# - datetime
# - typing
